// setfxweb/src/app/core/enums/basis.enum.ts
export enum Basis {
    M = "ACT/360",   
    D = "ACT+1/360",
    F = "ACT+1/365",
    J = "ACT/365 (JPY)",
    B = "ACT/365.25",
    N = "ACT/nACT",
    C = "30/360",
    I = "30E+1/360 (ITL)",
    A = "ACT/ACT",
    E = "30E/360",    
    Y = "30E/360 (FEB)",    
    Z = "ACT+1/365(Thai)",
    C2 = "BUS/252",
    C4 = "ACT/364",
    C5 = "ACT/365",
    C6 = "ACT/365 (366)"    
}